修正１

+++ app/controllers/items_controller.rb
@@ -21,8 +21,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user.id != item.user_id
+      redirect_to root_path
+    elsif
+     item.destroy
+     redirect_to root_path
+    end
   end
 
   def show
